sentrybot/requirements.txt
transformers
pillow
torch
distinctipy